Moved classes
Moved classes to be more structured.
