--- src/remotemapper/ConfigFiles.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (C) 2018 Robert Hutter
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package remotemapper;
-
-/**
- * A collection of configuration files for the application to run.
- * 
- * @author Robert Hutter
- */
-public enum ConfigFiles {
-    // Files
-    MAP("map.d"),
-    CMD_PRESETS("presets.xml");
-    
-    private final String file;
-    
-    ConfigFiles (String file)
-    {
-        this.file = file;
-    }
-    
-    public String getName ()
-    {
-        return file;
-    }
-}
